parser: move nextch() inside the parser closure and simplify.
* lisp.lua (nextch): Move inside parse, and use closure upvalues
of i and s instead of parameters.
(lex): Rename a local from n to j to avoid confusion over a
shadowed upvalue.
(M.parser): Avoid recalculating the length of s on every call by
setting it once as an upvalue when `M.parse (s)' is called,
massively speeding up parsing of long files.